More improvements to type interpretation.
Add TEXTUAL-CONVENTION decls to type defs. Interpret field types by
looking up the field identifiers in type defs directly instead of
trusting the declared type in the SEQUENCE. This is because the
top-level decls usually have better type info (for example BITS and
enumerations have bit names and named values at the top level but not in
the SEQUENCE).